Reorganize : Dedup code make package exporter
